(fix) defualt sandbox fuse for free execution

+++ src/lualib.h
@@ -46,7 +46,7 @@ LUAMOD_API int (luaopen_package) (lua_State *L);
 
 
 /* open all previous libraries */
-LUALIB_API void (luaL_openlibs) (lua_State *L ,int sandboxFuse);
+LUALIB_API void (luaL_openlibs) (lua_State *L ,long sandboxFuse);
 
 
 #endif
